refactor validations

+++ Backend/src/main/java/licenta/backend/helpers/UserRoomHelper.java
@@ -2,5 +2,5 @@ package licenta.backend.helpers;
 
 public interface UserRoomHelper
 {
-    int getroomid();
+long  getroomid();
 }
